hw07-patterns upd after review

+++ hw07-patterns/src/main/java/ru/otus/processor/homework/TimeProvider.java
@@ -2,5 +2,5 @@ package ru.otus.processor.homework;
 
 public interface TimeProvider {
 
-    int getSeconds();
+    long getSeconds();
 }
